requirements.txt
azure-storage-queue==12.3.0